refactor: rename `reporters` to `reporter`
